requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP client for AI integration
requests>=2.31.0

# Development and utility dependencies
python-dotenv>=1.0.0
pydantic>=2.4.0

# Optional: For enhanced logging and monitoring
rich>=13.0.0

# Optional: For better development experience
watchfiles>=0.21.0
